racingcar refactor: mvc패턴 적용
