+++ src/main/kotlin/racingcar/Main.kt
@@ -1,8 +1,10 @@
 package racingcar
 
-import racingcar.racing.Racing
-import racingcar.ui.InputView
+import racingcar.domain.car.Car
+import racingcar.domain.racing.Racing
+import racingcar.view.InputView
 import racingcar.util.Validation
+import racingcar.view.ResultView
 
 fun main() {
     val carsName = InputView.getCarsName().also {
@@ -11,6 +13,11 @@ fun main() {
     val attemptCount = InputView.getAttemptCount().also {
         Validation.checkTryCountCondition(it)
     }
-
-    Racing(carsName, attemptCount.toInt()).start()
+    val showRacingResult = { car: Car ->
+        ResultView.printRacingResult(car)
+    }
+    val showRacingWinner = { winners: List<String> ->
+        ResultView.printRacingWinner(winners)
+    }
+    Racing(carsName, attemptCount.toInt()).start(showRacingResult, showRacingWinner)
 }
