+++ src/main/kotlin/racingcar/domain/car/CarFactory.kt
@@ -1,6 +1,6 @@
-package racingcar.car
+package racingcar.domain.car
 
-import racingcar.engine.Engine
+import racingcar.domain.car.engine.Engine
 import racingcar.util.Validation
 
 object CarFactory {
