+++ src/main/kotlin/racingcar/domain/car/Cars.kt
@@ -1,14 +1,15 @@
-package racingcar.car
+package racingcar.domain.car
 
-import racingcar.engine.RandomEngine
+import racingcar.domain.car.engine.RandomEngine
 
 class Cars(carsName: String) {
     val carList: List<Car> = CarFactory.initCarList(carsName, RandomEngine())
 
     fun startRound(): List<Car> {
-        return carList.map { car ->
+        carList.forEach { car ->
             car.start()
         }
+        return carList
     }
 
     private fun getHighScore(): Int {
