+++ src/main/kotlin/racingcar/domain/car/CarsFactory.kt
@@ -1,4 +1,4 @@
-package racingcar.car
+package racingcar.domain.car
 
 object CarsFactory {
     fun create(carsName: String): Cars {
