+++ src/main/kotlin/racingcar/domain/car/engine/Engine.kt
@@ -1,4 +1,4 @@
-package racingcar.engine
+package racingcar.domain.car.engine
 
 interface Engine {
     fun forward(distance: Int): Int
