--- src/main/kotlin/racingcar/racing/Racing.kt
@@ -1,19 +1,0 @@
-package racingcar.racing
-
-import racingcar.car.CarsFactory
-import racingcar.ui.ResultView
-
-class Racing(carsName: String, private val attemptCount: Int) {
-    private val cars = CarsFactory.create(carsName)
-
-    fun start() {
-        repeat(attemptCount) {
-            cars.startRound().forEach { car ->
-                ResultView.printRacingResult(car)
-            }
-            println()
-        }
-
-        ResultView.printRacingWinner(cars.getWinner())
-    }
-}
