+++ src/test/kotlin/racingcar/CarFactoryTest.kt
@@ -3,7 +3,7 @@ package racingcar
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.car.Cars
+import racingcar.domain.car.Cars
 import racingcar.util.Validation
 
 class CarFactoryTest {
