+++ src/test/kotlin/racingcar/EngineTest.kt
@@ -2,8 +2,8 @@ package racingcar
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import racingcar.car.Car
-import racingcar.engine.GoEngine
+import racingcar.domain.car.Car
+import racingcar.domain.car.engine.GoEngine
 
 class EngineTest {
     @Test
